Clean up work after setting up Browserify and postcss
