+++ .eslintrc.js
@@ -7,7 +7,8 @@ module.exports = {
     "browser": true,
     "commonjs": true,
     "es6": true,
-    "node": true
+    "node": true,
+    "jquery": true
   },
   "extends": "eslint:recommended",
   "rules": {
