+++ grunt/aliases.yaml
@@ -3,9 +3,13 @@ js:
   - browserify
 css:
   - postcss
+html:
+  - copy:html
 build:
+  - clean
   - js
   - css
+  - html
 start:
   - build
   - connect
