+++ grunt/browserify.js
@@ -2,7 +2,7 @@ module.exports = {
   dist: {
     options: {
       transform: [
-        ['babelify', { presets: ['es2015', 'react'] }]
+        ['babelify', { presets: ['es2015'] }]
       ]
     },
     src: ['src/js/app.js'],
