+++ grunt/clean.js
@@ -1,0 +1,3 @@
+module.exports = {
+  clean: ['dist']
+};
