+++ grunt/copy.js
@@ -1,0 +1,8 @@
+module.exports = {
+  html: {
+    expand: true,
+    cwd: 'src',
+    src: ['{,**/}*.html'],
+    dest: 'dist/'
+  }
+};
