+++ grunt/watch.js
@@ -12,12 +12,14 @@ module.exports = {
   },
   css: {
     files: ['src/css/{,**/}*.css'],
+    tasks: ['css'],
     options: {
       livereload: true
     }
   },
   html: {
     files: ['src/{,**/}*.html'],
+    tasks: ['html'],
     options: {
       livereload: true
     }
