+++ package.json
@@ -5,6 +5,8 @@
   "description": "Implementing the Snake arcade game in Bacon.js",
   "main": "src/app.js",
   "scripts": {
+    "start": "./node_modules/grunt-cli/bin/grunt",
+    "build": "./node_modules/grunt-cli/bin/grunt build",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -34,10 +36,13 @@
     "grunt-browserify": "^4.0.1",
     "grunt-bump": "^0.7.0",
     "grunt-cli": "^0.1.13",
+    "grunt-contrib-clean": "^1.0.0",
     "grunt-contrib-connect": "^1.0.0",
+    "grunt-contrib-copy": "^1.0.0",
     "grunt-contrib-watch": "^0.6.1",
     "grunt-eslint": "^18.0.0",
     "grunt-git-committers": "^0.1.4",
+    "grunt-postcss": "^0.8.0",
     "grunt-templated-changelog": "^0.1.3",
     "load-grunt-config": "^0.19.1",
     "pixrem": "^3.0.0",
