+++ src/css/app.css
@@ -1,3 +1,7 @@
+body {
+  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
+}
+
 .row {
   height: 27px;
 }
