+++ src/index.html
@@ -1,8 +1,11 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>Implementing the Snake arcade game in Bacon.js</title>
-  <link rel="stylesheet" href="app.css" />
+  <link href="css/app.min.css" rel="stylesheet" />
 </head>
 <body>
 
@@ -11,9 +14,8 @@
 
   <div id="game"></div>
 
-  <script src="/node_modules/jquery/dist/jquery.min.js"></script>
-  <script src="/node_modules/baconjs/dist/Bacon.min.js"></script>
-  <script src="app.js"></script>
+  <script src="../node_modules/jquery/dist/jquery.min.js"></script>
+  <script src="js/bundle.js"></script>
 
 </body>
 </html>
